# Optimize service sorting