--- a/pilot/pkg/model/push_context.go
+++ b/pilot/pkg/model/push_context.go
@@ -1615,16 +1615,15 @@ func resolveServiceAliases(allServices []*Service, configsUpdated sets.Set[Confi
 // SortServicesByCreationTime sorts the list of services in ascending order by their creation time (if available).
 func SortServicesByCreationTime(services []*Service) []*Service {
 	slices.SortStableFunc(services, func(a, b *Service) int {
-		r := a.CreationTime.Compare(b.CreationTime)
+		if r := a.CreationTime.Compare(b.CreationTime); r != 0 {
+			return r
+		}
 		// If creation time is the same, then behavior is nondeterministic. In this case, we can
 		// pick an arbitrary but consistent ordering based on name and namespace, which is unique.
 		// CreationTimestamp is stored in seconds, so this is not uncommon.
-		if r == 0 {
-			an := a.Attributes.Name + "." + a.Attributes.Namespace
-			bn := b.Attributes.Name + "." + b.Attributes.Namespace
-			return cmp.Compare(an, bn)
-		}
-		return r
+		an := a.Attributes.Name + "." + a.Attributes.Namespace
+		bn := b.Attributes.Name + "." + b.Attributes.Namespace
+		return cmp.Compare(an, bn)
 	})
 	return services
 }