Changed `sort.SliceStable` to `slices.SortStableFunc`.

```
goos: darwin
goarch: arm64
pkg: istio.io/istio/pilot/pkg/model
                             │   old.txt   │              new.txt               │
                             │   sec/op    │   sec/op     vs base               │
SortServicesByCreationTime-8   8.795µ ± 2%   5.005µ ± 1%  -43.10% (p=0.002 n=6)

                             │  old.txt   │              new.txt              │
                             │    B/op    │   B/op     vs base                │
SortServicesByCreationTime-8   56.00 ± 0%   0.00 ± 0%  -100.00% (p=0.002 n=6)

                             │  old.txt   │              new.txt               │
                             │ allocs/op  │ allocs/op   vs base                │
SortServicesByCreationTime-8   2.000 ± 0%   0.000 ± 0%  -100.00% (p=0.002 n=6)
```